fix(client_cli): `disconnect` broken from socket
When a connection with a server is broken by the server, the CLI side
socket fails to return a remote endpoint. This caused the `disconnect`
command to fail before it could reset local connection state, thus
keeping the CLI in an unrecoverable state where it is unable to connect
or disconnect.